backend/requirements.txt
attrs==22.1.0
click==8.1.3
et-xmlfile==1.1.0
Flask==2.2.2
gunicorn==20.1.0
importlib-metadata==4.12.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.2
opencv-python==4.6.0.66
openpyxl==3.0.10
packaging==21.3
Pillow==9.2.0
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
pyparsing==3.0.9
pytest==7.1.2
stripe==4.2.0
tomli==2.0.1
Werkzeug==2.2.2
zipp==3.8.1